Nouvelle  architecture conversation_service#4
