--- api/websocket.py
@@ -1,14 +1,0 @@
-from fastapi import APIRouter, WebSocket
-from starlette.websockets import WebSocketDisconnect
-
-router = APIRouter()
-
-@router.websocket("/ws")
-async def websocket_endpoint(ws: WebSocket):
-    await ws.accept()
-    try:
-        while True:
-            data = await ws.receive_text()
-            await ws.send_text(data)
-    except WebSocketDisconnect:
-        pass
