--- conversation_service/clients/__init__.py
@@ -1,7 +1,0 @@
-"""Asynchronous service clients used by the conversation service."""
-
-from .openai_client import OpenAIClient
-from .search_client import SearchClient
-from .cache_client import CacheClient
-
-__all__ = ["OpenAIClient", "SearchClient", "CacheClient"]
