+++ conversation_service/core/errors.py
