+++ conversation_service/core/helpers.py
