--- conversation_service/utils/__init__.py
@@ -1,13 +1,0 @@
-"""Utility helpers for the conversation service."""
-
-from .logging import get_structured_logger
-from .decorators import metrics, cache
-from .helpers import chunks, flatten_dict
-
-__all__ = [
-    "get_structured_logger",
-    "metrics",
-    "cache",
-    "chunks",
-    "flatten_dict",
-]
