--- core/__init__.py
@@ -1,25 +1,0 @@
-"""Core utilities shared across the project."""
-
-from .logging import get_logger, setup_logging, JsonFormatter
-from .decorators import metrics, cache
-from .helpers import chunks, flatten_dict
-from .validators import non_empty_str, positive_number, percentage
-
-__all__ = [
-    "get_logger",
-    "setup_logging",
-    "JsonFormatter",
-    "metrics",
-    "cache",
-    "chunks",
-    "flatten_dict",
-    "non_empty_str",
-    "positive_number",
-    "percentage",
-]
-"""Core utilities for shared validation logic and pipeline execution."""
-
-from .agent_pipeline import AgentPipeline, PipelineStep
-from .fallback_manager import FallbackManager
-
-__all__ = ["AgentPipeline", "PipelineStep", "FallbackManager"]
